Extract Component API to separate file
