+++ lib/react/component/api.rb
@@ -1,0 +1,47 @@
+module React
+  module Component
+    module API
+      include Native
+
+      alias_native :dom_node, :getDOMNode
+      alias_native :mounted?, :isMounted
+      alias_native :force_update!, :forceUpdate
+
+      def set_props(prop, &block)
+        raise "No native ReactComponent associated" unless @native
+        %x{
+          #{@native}.setProps(#{prop.shallow_to_n}, function(){
+            #{block.call if block}
+          });
+        }
+      end
+
+      def set_props!(prop, &block)
+        raise "No native ReactComponent associated" unless @native
+        %x{
+          #{@native}.replaceProps(#{prop.shallow_to_n}, function(){
+            #{block.call if block}
+          });
+        }
+      end
+
+      def set_state(state, &block)
+        raise "No native ReactComponent associated" unless @native
+        %x{
+          #{@native}.setState(#{state.shallow_to_n}, function(){
+            #{block.call if block}
+          });
+        }
+      end
+
+      def set_state!(state, &block)
+        raise "No native ReactComponent associated" unless @native
+        %x{
+          #{@native}.replaceState(#{state.shallow_to_n}, function(){
+            #{block.call if block}
+          });
+        }
+      end
+    end
+  end
+end
